requirements.txt
Django==1.3.1
MySQL-python==1.2.2
PIL==1.1.7
South==0.7
django-admin-tools==0.4.1
django-batch-select==0.2.4
django-debug-toolbar==0.9.4
django-mptt
django-pagination==1.0.7
django-social-auth==0.7.6
html5lib==0.95
httplib2==0.7.2
oauth==1.0a
oauth2==1.5.211
psycopg2==2.4.4
pyPdf==1.13
pymorphy==0.5.6
pytils==0.2.3
reportlab==2.5
sorl-thumbnail==11.12
urllib2-file==0.2.1
xhtml2pdf==0.0.4
